Remove usages of ?selectableItemBackground, it isn't needed - this may end up fixing #1629
