+++ core/src/main/java/com/afollestad/materialdialogs/utils/DialogExt.kt
@@ -26,6 +26,8 @@ import com.afollestad.materialdialogs.R
 import com.afollestad.materialdialogs.callbacks.invokeAll
 import com.afollestad.materialdialogs.checkbox.getCheckBoxPrompt
 import com.afollestad.materialdialogs.customview.CUSTOM_VIEW_NO_PADDING
+import com.afollestad.materialdialogs.utils.MDUtil.getDrawable
+import com.afollestad.materialdialogs.utils.MDUtil.getString
 
 internal fun MaterialDialog.setWindowConstraints() {
   window!!.setSoftInputMode(SOFT_INPUT_ADJUST_RESIZE)
@@ -118,7 +120,7 @@ internal fun MaterialDialog.populateText(
   typeface: Typeface?,
   textColor: Int? = null
 ) {
-  val value = text ?: Util.getString(this, textRes, fallback)
+  val value = text ?: getString(this, textRes, fallback)
   if (value != null) {
     (textView.parent as View).visibility = View.VISIBLE
     textView.visibility = View.VISIBLE
