+++ files/src/main/res/layout/md_file_chooser_item.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:background="?selectableItemBackground"
     android:orientation="vertical"
     android:paddingBottom="@dimen/file_chooser_item_vertical_padding"
     android:paddingLeft="@dimen/file_chooser_item_horizontal_padding"
